Move test method list into common package
